Remove mbuffer support
The idea was for mbuffer to be faster at reading than ruby can be.
Unfortunately, mbuffer splits its memory buffer into blocks and forwards
data only when at least one block was read. Since these blocks are
rather large, it adds delay to log processing. mbuffer integration was
never in use for this purpose, so remove it altogether.